Reduced OrderdDict-Abuse
... Where do I even start? This is NOT how you should use OrderedDicts. WHY would you make lists of keys of OrderedDict of lists of strings of strings of lists? The effect is the same if you use the lists directly, except the latter is much faster, easier to read, etc. Why use OrderedDict for simple layout-design? 22 occurrences removed...

And I need to double-check envi_mat.py, but why would you keep a dictionary-copy of data you store as OrderedDict as well? It only makes sense if you allow changing one of the copies... which we dont? Frankly, it seems, that of all 50+ occurrences of OrderedDict, maybe ten are necessary.