Idea for OrderedDict
OrderedDicts could be collapsed to all dicts of type ...d to be the same as the ordered dicts (e.g. self.metal_datd = self.metal_dat). The only reason why I didn't already completely rewrite this, is because I don't know if there is a hidden way to modify the dicts. If so, we would need a deepcopy...