Minor Cleanups
Partially sorting imports.
TODO: Why is every socket and node imported individually? Same for the operators, etc. ? Isn't it better to use sub-(un)register functions? Or if you really need to register them here, why not import them as lists, to keep it clean? ... The only reason I could think of is registration order which A) would be a sign of poor module design and B) could be fixed here anyways...