"Shading Divices", minimally invasive, first pitch
"Shading devices" is a minimally invasive subsystem providing support for creation/simulation of shading devices from within the VI-Suite.
Note:
We will try to keep API-compatibility and file-compatibility as long as possible. Therefore expect untidy code and weird stunts.
Background:
In 2016 Felix Loehr wrote multiple updates and patches for the VI-Suite, in order to add support for energy plus shading devices. They also enhanced the interface, the speed and the robustness against some edgecase-problems, all while maintaining compatibility of the simulation files (.blend-files).
Sadly even though presented to the creator of the VI-Suite, and him even openly admiring the functionality and code quality, the patches were never applied to the main repo.
After having to re-patch newer versions of the vi-suite, in december 2017 Loehr separated the core functionality (shading devices) as far as possible for ease of maintenance and porting. While lacking former elegance and speedups, the new code only touched the original vi-suite in 8 places (2 files).
This meant having to play a few dirty tricks (like overriding / shadowing classes on import), but allowed (de)activation by (un)commenting two lines (in the init file) and even to write a new addon that (un)patched the vi-suite accordingly when (de)activated.
This code is the basis for patching this fork of the vi-suite.