Repair GitHubs F-ups
Somehow git f-ed up in at least ONE of the recent commits. Instead of only commiting my changes, it also managed to swap lines, which in turn made the VI-Suite ill-defined. Luckily it was easily found as a syntax-error, but other errors could very well be possible... (facepalm)
The missing ":" in vi_node however was my mistake in commit 73cedbd1 (Minor cleanups on vi_node).