rewrote path_update
The function "path_update" now resides in "paths.py" and is much more powerfull. It may read absolute or blend-file relative paths, or paths starting with a ".". These latter ones, it may find relative to the blender-binary, the python-binary, the addons-folders, the script-paths, the user-ressources or the system-ressources... If the path isn't found or not a dir, or was ill-defined, a helpfull message is print to the console and the path is skipped... as are empty paths.